requirements.txt
Flask==2.2.3
requests==2.32.3
python-dotenv==1.0.0
openai==1.2.0
together==1.2.12
werkzeug==2.2.3